requirements.txt
absl-py==0.3.0
astor==0.7.1
audioread==2.1.6
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
joblib==0.12.2
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
librosa==0.6.2
llvmlite==0.24.0
Markdown==2.6.11
numba==0.39.0
numpy==1.14.5
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pushbullet.py==0.11.0
Pygments==2.2.0
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
PyYAML==3.13
requests==2.19.1
resampy==0.2.1
scikit-learn==0.19.2
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
tqdm==4.24.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
